# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface FileUpload {
  file: Blob;
}

export interface PriceList {
  id: number;
  name: string;
  /** @nullable */
  organization: string | null;
  /** @nullable */
  year: number | null;
  itemCount: number;
  createdAt: string;
}

export interface PriceListInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  organization?: string | null;
  /** @nullable */
  year?: number | null;
}

export interface PriceListUploadResult {
  priceList: PriceList;
  added: number;
  updated: number;
  skipped: number;
  warnings: string[];
}

export interface PriceItem {
  id: number;
  priceListId: number;
  pozNo: string;
  description: string;
  /** @nullable */
  unit: string | null;
  unitPrice: number;
  createdAt: string;
}

export interface PriceItemInput {
  priceListId: number;
  /** @minLength 1 */
  pozNo: string;
  /** @minLength 1 */
  description: string;
  /** @nullable */
  unit?: string | null;
  unitPrice: number;
}

export interface PriceItemUpdate {
  /** @minLength 1 */
  pozNo?: string;
  /** @minLength 1 */
  description?: string;
  /** @nullable */
  unit?: string | null;
  unitPrice?: number;
}

export interface TenderProject {
  id: number;
  name: string;
  /** @nullable */
  fileName: string | null;
  itemCount: number;
  matchedCount: number;
  fuzzyCount: number;
  unmatchedCount: number;
  /** @nullable */
  totalAmount: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface TenderProjectInput {
  /** @minLength 1 */
  name: string;
}

export type TenderItemMatchStatus = typeof TenderItemMatchStatus[keyof typeof TenderItemMatchStatus];


export const TenderItemMatchStatus = {
  matched: 'matched',
  fuzzy: 'fuzzy',
  unmatched: 'unmatched',
  manual: 'manual',
} as const;

export interface TenderItem {
  id: number;
  tenderProjectId: number;
  rowOrder: number;
  pozNo: string;
  description: string;
  /** @nullable */
  unit: string | null;
  quantity: number;
  /** @nullable */
  unitPrice: number | null;
  /** @nullable */
  totalPrice: number | null;
  matchStatus: TenderItemMatchStatus;
  /** @nullable */
  matchScore: number | null;
  /** @nullable */
  matchedPriceItemId: number | null;
  /** @nullable */
  matchedPozNo: string | null;
  /** @nullable */
  matchedDescription: string | null;
  /** @nullable */
  matchedSourceName: string | null;
  createdAt: string;
}

export interface TenderUploadResult {
  project: TenderProject;
  items: TenderItem[];
  warnings: string[];
}

export interface TenderItemUpdate {
  /** @nullable */
  unitPrice?: number | null;
  /** @nullable */
  matchedPriceItemId?: number | null;
}

export interface MatchRequest {
  /**
     * Eşleştirme için kullanılacak liste id'leri. Boş/atlanırsa tüm listeler kullanılır.
     * @nullable
     */
  priceListIds?: number[] | null;
}

export type ListPriceItemsParams = {
/**
 * Belirtilirse sadece bu listedeki kalemler döner. Atlanırsa tüm listelerde arama yapılır.
 */
priceListId?: number;
/**
 * Poz no veya tanım içinde metin araması.
 */
q?: string;
limit?: number;
};