# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /price-lists:
    get:
      operationId: listPriceLists
      tags: [priceLists]
      summary: Kayıtlı birim fiyat listelerini getir
      responses:
        "200":
          description: Birim fiyat listeleri
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceList"
    post:
      operationId: createPriceList
      tags: [priceLists]
      summary: Yeni birim fiyat listesi kaynağı oluştur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceListInput"
      responses:
        "201":
          description: Oluşturuldu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceList"
        "400":
          description: Geçersiz istek
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /price-lists/{id}:
    get:
      operationId: getPriceList
      tags: [priceLists]
      summary: Birim fiyat listesi detayını getir
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Birim fiyat listesi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceList"
        "404":
          description: Bulunamadı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePriceList
      tags: [priceLists]
      summary: Birim fiyat listesini ve kalemlerini sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Silindi
        "404":
          description: Bulunamadı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /price-lists/{id}/upload:
    post:
      operationId: uploadPriceListFile
      tags: [priceLists]
      summary: Excel veya PDF birim fiyat dosyası yükleyip poz kalemlerini ayrıştır
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/FileUpload"
      responses:
        "200":
          description: Ayrıştırma sonucu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceListUploadResult"
        "400":
          description: Dosya ayrıştırılamadı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Bulunamadı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /price-items:
    get:
      operationId: listPriceItems
      tags: [priceLists]
      summary: Poz kalemlerini getir (isteğe bağlı liste filtresi ve arama ile)
      parameters:
        - name: priceListId
          in: query
          required: false
          schema:
            type: integer
          description: Belirtilirse sadece bu listedeki kalemler döner. Atlanırsa tüm listelerde arama yapılır.
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Poz no veya tanım içinde metin araması.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Poz kalemleri
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceItem"
    post:
      operationId: createPriceItem
      tags: [priceLists]
      summary: Bir listeye manuel poz kalemi ekle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceItemInput"
      responses:
        "201":
          description: Oluşturuldu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceItem"
        "400":
          description: Geçersiz istek
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /price-items/{id}:
    patch:
      operationId: updatePriceItem
      tags: [priceLists]
      summary: Poz kalemini güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceItemUpdate"
      responses:
        "200":
          description: Güncellendi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceItem"
        "404":
          description: Bulunamadı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePriceItem
      tags: [priceLists]
      summary: Poz kalemini sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Silindi
        "404":
          description: Bulunamadı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /tender-items/{id}:
    patch:
      operationId: updateTenderItem
      tags: [tenderProjects]
      summary: Kalemin birim fiyatını elle düzenle veya belirli bir poz eşleşmesi seç
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenderItemUpdate"
      responses:
        "200":
          description: Güncellendi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenderItem"
        "404":
          description: Bulunamadı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    FileUpload:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
    PriceList:
      type: object
      required: [id, name, organization, year, itemCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        organization:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        itemCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    PriceListInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        organization:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
    PriceListUploadResult:
      type: object
      required: [priceList, added, updated, skipped, warnings]
      properties:
        priceList:
          $ref: "#/components/schemas/PriceList"
        added:
          type: integer
        updated:
          type: integer
        skipped:
          type: integer
        warnings:
          type: array
          items:
            type: string
    PriceItem:
      type: object
      required: [id, priceListId, pozNo, description, unit, unitPrice, createdAt]
      properties:
        id:
          type: integer
        priceListId:
          type: integer
        pozNo:
          type: string
        description:
          type: string
        unit:
          type: ["string", "null"]
        unitPrice:
          type: number
        createdAt:
          type: string
          format: date-time
    PriceItemInput:
      type: object
      required: [priceListId, pozNo, description, unitPrice]
      properties:
        priceListId:
          type: integer
        pozNo:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        unit:
          type: ["string", "null"]
        unitPrice:
          type: number
    PriceItemUpdate:
      type: object
      properties:
        pozNo:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        unit:
          type: ["string", "null"]
        unitPrice:
          type: number
...
    TenderItem:
      type: object
      required:
        [
          id,
          tenderProjectId,
          rowOrder,
          pozNo,
          description,
          unit,
          quantity,
          unitPrice,
          totalPrice,
          matchStatus,
          matchScore,
          matchedPriceItemId,
          matchedPozNo,
          matchedDescription,
          matchedSourceName,
          createdAt,
        ]
      properties:
        id:
          type: integer
        tenderProjectId:
          type: integer
        rowOrder:
          type: integer
        pozNo:
          type: string
        description:
          type: string
        unit:
          type: ["string", "null"]
        quantity:
          type: number
        unitPrice:
          type: ["number", "null"]
        totalPrice:
          type: ["number", "null"]
        matchStatus:
          type: string
          enum: [matched, fuzzy, unmatched, manual]
        matchScore:
          type: ["number", "null"]
        matchedPriceItemId:
          type: ["integer", "null"]
        matchedPozNo:
          type: ["string", "null"]
        matchedDescription:
          type: ["string", "null"]
        matchedSourceName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    TenderItemUpdate:
      type: object
      properties:
        unitPrice:
          type: ["number", "null"]
        matchedPriceItemId:
          type: ["integer", "null"]
...