# lib/api-zod/src/generated/types/priceItemInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PriceItemInput {
  priceListId: number;
  /** @minLength 1 */
  pozNo: string;
  /** @minLength 1 */
  description: string;
  /** @nullable */
  unit?: string | null;
  unitPrice: number;
}