# lib/api-zod/src/generated/types/tenderItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TenderItemMatchStatus } from './tenderItemMatchStatus';

export interface TenderItem {
  id: number;
  tenderProjectId: number;
  rowOrder: number;
  pozNo: string;
  description: string;
  /** @nullable */
  unit: string | null;
  quantity: number;
  /** @nullable */
  unitPrice: number | null;
  /** @nullable */
  totalPrice: number | null;
  matchStatus: TenderItemMatchStatus;
  /** @nullable */
  matchScore: number | null;
  /** @nullable */
  matchedPriceItemId: number | null;
  /** @nullable */
  matchedPozNo: string | null;
  /** @nullable */
  matchedDescription: string | null;
  /** @nullable */
  matchedSourceName: string | null;
  createdAt: Date;
}